+++ feature/src/main/java/com/zcard/feature/cardeditor/transform/TransformSideEffect.kt
@@ -1,4 +1,4 @@
-package com.zcard.feature.cardeditor
+package com.zcard.feature.cardeditor.transform
 
 import androidx.annotation.StringRes
